refactored addon service and fixed stream search in account-profile service use-cases and so on
