+++ apps/account-profile-service/app/adapters/addon_provider.py
