+++ apps/addon-controller-service/app/api/v1/routes.py
@@ -1,42 +1,38 @@
 from dependency_injector.wiring import inject, Provide
-from fastapi import APIRouter, Body, Depends, Query
+from fastapi import APIRouter, Depends, Query
 from fastapi_factory.responses import create_success_response, SuccessResponse
 
 from ...containers import Container
-from ...services.services import IAddonService
+from .models import CatalogRequestModel, MetaRequestModel
 from core.pydantic.addons.manifest import AddonManifest
-from core.pydantic.catalog.catalog import CatalogResponse, CatalogRequest
+from core.pydantic.catalog.catalog import CatalogResponse
 from core.pydantic.meta.meta import MetaResponse
-from ...services.services import IAddonService
-from pydantic import BaseModel, Field
-from typing import Optional
+
+# Import Use Cases
+from ...use_cases.get_manifest import GetManifestUseCase
+from ...use_cases.get_catalog import GetCatalogUseCase
+from ...use_cases.get_meta import GetMetaUseCase
 
 router = APIRouter()
-
-
-class MetaRequest(BaseModel):
-    manifest_url: str = Field(..., alias="manifestUrl")
-    item_type: Optional[str] = Field(None, alias="itemType")
 
 
 @router.get("/manifest", response_model=SuccessResponse[AddonManifest])
 @inject
 async def get_addon_manifest(
     url: str = Query(..., description="The URL of the addon's manifest.json"),
-    addon_service: IAddonService = Depends(Provide[Container.addon_service]),
+    use_case: GetManifestUseCase = Depends(Provide[Container.get_manifest_use_case]),
 ):
-    manifest = await addon_service.get_manifest(url)
+    manifest = await use_case.execute(url)
     return create_success_response(data=manifest)
 
 
 @router.post("/catalog", response_model=SuccessResponse[CatalogResponse])
 @inject
 async def get_addon_catalog(
-    request: CatalogRequest,
-    addon_service: IAddonService = Depends(Provide[Container.addon_service]),
+    request: CatalogRequestModel,
+    use_case: GetCatalogUseCase = Depends(Provide[Container.get_catalog_use_case]),
 ):
-    """[INTERNAL] Fetches a content catalog from an addon's manifest URL."""
-    catalog = await addon_service.get_catalog(
+    catalog = await use_case.execute(
         manifest_url=request.manifest_url,
         catalog_type=request.catalog_type,
         catalog_id=request.catalog_id,
@@ -49,13 +45,12 @@ async def get_addon_catalog(
 @inject
 async def get_addon_meta(
     item_id: str,
-    request: MetaRequest,  # USE THE NEW MODEL
-    addon_service: IAddonService = Depends(Provide[Container.addon_service]),
+    request: MetaRequestModel,
+    use_case: GetMetaUseCase = Depends(Provide[Container.get_meta_use_case]),
 ):
-    """[INTERNAL] Fetches detailed metadata for an item from an addon."""
-    meta = await addon_service.get_meta(
+    meta = await use_case.execute(
         manifest_url=request.manifest_url,
         item_id=item_id,
-        item_type=request.item_type,  # PASS THE TYPE
+        item_type=request.item_type,
     )
     return create_success_response(data=meta)
